+++ src/spaces/grid_single.jl
@@ -11,6 +11,7 @@ export GridSpaceSingle, id_in_position
 struct GridSpaceSingle{D,P} <: AbstractGridSpace{D,P}
     stored_ids::Array{Int,D}
     metric::Symbol
+    offsets_at_radius::Dict{Int,Vector{NTuple{D,Int}}}
     offsets_within_radius::Dict{Float64,Vector{NTuple{D,Int}}}
     offsets_within_radius_no_0::Dict{Float64,Vector{NTuple{D,Int}}}
 end
@@ -29,7 +30,9 @@ All arguments and keywords behave exactly as in [`GridSpace`](@ref).
 function GridSpaceSingle(d::NTuple{D,Int}; periodic = true, metric = :chebyshev) where {D}
     s = zeros(Int, d)
     return GridSpaceSingle{D,periodic}(s, metric,
-        Dict{Float64,Vector{NTuple{D,Int}}}(), Dict{Float64,Vector{NTuple{D,Int}}}(),
+        Dict{Int,Vector{NTuple{D,Int}}}(),
+        Dict{Float64,Vector{NTuple{D,Int}}}(),
+        Dict{Float64,Vector{NTuple{D,Int}}}(),
     )
 end
 # Implementation of space API