+++ test/grid_space_tests.jl
@@ -1,11 +1,6 @@
 using Test, Agents, Random
 using StableRNGs
 
-mutable struct GridAgent2D <: AbstractAgent
-    id::Int
-    pos::Dims{2}
-end
-
 @testset "$(SpaceType)" for SpaceType in (GridSpace, GridSpaceSingle)
     @testset "size, dim=$D" for D in (1,3)
         dims = (fill(5, D)...,)
@@ -20,7 +15,7 @@ end
 
     @testset "add/move/kill" begin
         space = SpaceType((3, 3))
-        model = ABM(GridAgent2D, space; rng = StableRNG(42))
+        model = ABM(GridAgent{2}, space; rng = StableRNG(42))
         pos0 = (2,2)
         agent = add_agent!(pos0, model)
         id0 = agent.id
@@ -46,7 +41,7 @@ end
 
     @testset "positions + empty" begin
         space = SpaceType((3, 3))
-        model = ABM(GridAgent2D, space)
+        model = ABM(GridAgent{2}, space)
         empty = collect(empty_positions(model))
         @test length(empty) == 9
         locs_to_add = [1, 2, 3, 4, 5, 6, 9]
@@ -85,23 +80,23 @@ end
 
     @testset "Distances" begin
     @testset "Euclidean distance" begin
-        model = ABM(GridAgent2D, SpaceType((12, 10); periodic = true))
+        model = ABM(GridAgent{2}, SpaceType((12, 10); periodic = true))
         a = add_agent!((1.0, 6.0), model)
         b = add_agent!((11.0, 4.0), model)
         @test euclidean_distance(a, b, model) ≈ 2.82842712
 
-        model = ABM(GridAgent2D, SpaceType((12, 10); periodic = false))
+        model = ABM(GridAgent{2}, SpaceType((12, 10); periodic = false))
         a = add_agent!((1.0, 6.0), model)
         b = add_agent!((11.0, 4.0), model)
         @test euclidean_distance(a, b, model) ≈ 10.198039
     end
     @testset "Manhattan Distance" begin
-        model = ABM(GridAgent2D, SpaceType((12, 10); metric = :manhattan, periodic = true))
+        model = ABM(GridAgent{2}, SpaceType((12, 10); metric = :manhattan, periodic = true))
         a = add_agent!((1.0, 6.0), model)
         b = add_agent!((11.0, 4.0), model)
         @test manhattan_distance(a, b, model) ≈ 4
 
-        model = ABM(GridAgent2D, SpaceType((12, 10); metric = :manhattan, periodic = false))
+        model = ABM(GridAgent{2}, SpaceType((12, 10); metric = :manhattan, periodic = false))
         a = add_agent!((1.0, 6.0), model)
         b = add_agent!((11.0, 4.0), model)
         @test manhattan_distance(a, b, model) ≈ 12
@@ -117,7 +112,7 @@ end
             @testset "Metric=$(metric)" for metric in metrics
                 # To undersatnd where the numbers here are coming from,
                 # check out the plot in the docs that shows the metrics
-                model = ABM(GridAgent2D, SpaceType((5, 5); metric, periodic))
+                model = ABM(GridAgent{2}, SpaceType((5, 5); metric, periodic))
                 if metric ∈ (:euclidean, :mahnattan) # for r = 1 they give the same
                     @test sort!(collect(nearby_positions((2, 2), model))) ==
                         sort!([(2, 1), (1, 2), (3, 2), (2, 3)])
@@ -164,7 +159,7 @@ end
 
             end
             # also test larger r. See figure at docs for metrics to get the numbers
-            models = [ABM(GridAgent2D, SpaceType((9, 9); metric, periodic)) for metric in metrics]
+            models = [ABM(GridAgent{2}, SpaceType((9, 9); metric, periodic)) for metric in metrics]
             for m in models; fill_space!(m); end
             near_pos = [collect(nearby_positions((5,5), m, 3.4)) for m in models]
             near_ids = [collect(nearby_ids((5,5), m, 3.4)) for m in models] # this is 1 more
@@ -178,7 +173,7 @@ end
 
     @testset "Random nearby" begin
         # Test random_nearby_*
-        abm = ABM(GridAgent2D, GridSpace((10, 10)); rng = StableRNG(42))
+        abm = ABM(GridAgent{2}, GridSpace((10, 10)); rng = StableRNG(42))
         fill_space!(abm)
 
         nearby_id = random_nearby_id(abm[1], abm, 5)
@@ -199,7 +194,7 @@ end
 
     @testset "walk!" begin
         # Periodic
-        model = ABM(GridAgent2D, GridSpace((3, 3); periodic = true))
+        model = ABM(GridAgent{2}, GridSpace((3, 3); periodic = true))
         a = add_agent!((1, 1), model)
         walk!(a, (0, 1), model) # North
         @test a.pos == (1, 2)
@@ -227,7 +222,7 @@ end
         walk!(a, (1, 0), model; ifempty = true)
         @test a.pos == (2, 1)
         # aperiodic
-        model = ABM(GridAgent2D, GridSpace((3, 3); periodic = false))
+        model = ABM(GridAgent{2}, GridSpace((3, 3); periodic = false))
         a = add_agent!((1, 1), model)
         walk!(a, (0, 1), model) # North
         @test a.pos == (1, 2)
@@ -242,31 +237,92 @@ end
         walk!(a, (-1, -1), model) # Boundary in one direction, not in the other, attempt South west
         @test a.pos == (1, 2)
         @test_throws MethodError walk!(a, (1.0, 1.5), model) # Must use Int for gridspace
-        # Random Walks
-        model = ABM(GridAgent2D, GridSpace((5, 5)); rng = StableRNG(65))
-        a = add_agent!((3, 3), model)
-        walk!(a, rand, model)
-        @test a.pos == (4, 2)
-        walk!(a, rand, model)
-        @test a.pos == (5, 1)
-        walk!(a, rand, model)
-        @test a.pos == (4, 5)
-        
+
         # GridSpaceSingle
-        model = ABM(GridAgent2D, GridSpaceSingle((5, 5)))
+        model = ABM(GridAgent{2}, GridSpaceSingle((5, 5)))
         a = add_agent!((3, 3), model)
         walk!(a, (1, 1), model)
         a.pos == (4, 4)
 
         # Just a single sanity test for higher dimensions, just in case
-        mutable struct GridAgent3D <: AbstractAgent
-            id::Int
-            pos::Dims{3}
-        end
-        model = ABM(GridAgent3D, GridSpace((3, 3, 2)))
+        model = ABM(GridAgent{3}, GridSpace((3, 3, 2)))
         a = add_agent!((1, 1, 1), model)
         walk!(a, (1, 1, 1), model)
         @test a.pos == (2, 2, 2)
+    end
+
+    @testset "random walk" begin
+        # random walks on grid spaces with euclidean metric are not defined
+        space = SpaceType((10,10), metric=:euclidean)
+        model = ABM(GridAgent{2}, space)
+        add_agent!(model)
+        r = 1.0
+        @test_throws ArgumentError randomwalk!(model.agents[1], model, r)
+
+        # chebyshev metric
+        space = SpaceType((100,100), metric=:chebyshev)
+        model = ABM(GridAgent{2}, space)
+        x₀ = (50,50)
+        add_agent!(x₀, model)
+        r = 1.5
+        randomwalk!(model.agents[1], model, r)
+        x₁ = model.agents[1].pos
+        # chebyshev distance after the random step should be 1
+        @test maximum(abs.(x₁ .- x₀)) == 1
+        # for r < 1 the agent should not move
+        r = 0.5
+        randomwalk!(model.agents[1], model, r)
+        @test model.agents[1].pos == x₁
+
+        # manhattan metric
+        space = SpaceType((100,100), metric=:manhattan)
+        model = ABM(GridAgent{2}, space)
+        x₀ = (50,50)
+        add_agent!(x₀, model)
+        r = 1.5
+        randomwalk!(model.agents[1], model, r)
+        x₁ = model.agents[1].pos
+        # manhattan distance after the random step should be 1
+        @test manhattan_distance(x₁, x₀, model) == 1
+        # for r < 1 the agent should not move
+        r = 0.5
+        randomwalk!(model.agents[1], model, r)
+        @test model.agents[1].pos == x₁
+
+        space = SpaceType((100,100), metric=:manhattan)
+        model = ABM(GridAgent{2}, space)
+        pos = (50,50)
+        add_agent!(pos, model) # agent id = 1
+        # fill surrounding positions with other agents
+        offsets = [(-1,0), (1,0), (0,1), (0,-1)]
+        for β in offsets; add_agent!(pos.+β, model); end
+        if SpaceType == GridSpaceSingle
+            # agent 1 should not move since there are no available offsets
+            randomwalk!(model.agents[1], model, 1)
+            @test model.agents[1].pos == pos
+            # the keyword ifempty should have no effect in a GridSpaceSingle
+            randomwalk!(model.agents[1], model, 1; ifempty=false)
+            @test model.agents[1].pos == pos
+            randomwalk!(model.agents[1], model, 1; ifempty=true)
+            @test model.agents[1].pos == pos
+            # if agent 2 (49,50) moves, then agent 1 can only move
+            # to the position that was just freed
+            randomwalk!(model.agents[2], model, 1)
+            randomwalk!(model.agents[1], model, 1)
+            @test model.agents[1].pos == (49,50)
+        elseif SpaceType == GridSpace
+            # if ifempty=true (default), agent 1 should not move since there are
+            # no available offsets
+            randomwalk!(model.agents[1], model, 1)
+            @test model.agents[1].pos == pos
+            # if ifempty=false, agent 1 will move and occupy
+            # the same position as one of the other agents
+            randomwalk!(model.agents[1], model, 1; ifempty=false)
+            @test model.agents[1].pos ≠ pos
+            # 5 agents but only 4 unique positions
+            unique_pos = unique([a.second.pos for a in model.agents])
+            @test (length(model.agents)==5) && length(unique_pos)==4
+        end
     end
 
 end